Add unit test coverage for basic search handler Options methods.
